Work around a bug in composer that selects a feature branch of Robo.
We do not directly depend on Robo, but adding this line ensures that if we
update dependencies using the --prefer-lowest flag then Composer will
not accidentally select the `states` feature branch. It is likely that
this bug is caused by the presence of the alias "dev-state": "1.x-dev" in
the composer manifest of Robo.